% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_fsvm.R
\name{as_fsvm}
\alias{as_fsvm}
\title{Format field data for machine learning analysis}
\usage{
as_fsvm(
  genus = Model$Genus[i],
  species = Model$Species[i],
  subspecies = Model$SubSpecies[i],
  fielddata = fielddata,
  DataType = "DataType",
  Prcnt_C = "Prcnt_C"
)
}
\arguments{
\item{genus}{A character string corresponding with a single plant genus}

\item{species}{A character string corresponding with a single plant specific
epithet}

\item{subspecies}{A character string corresponding with a single plant
infraspecific eptithet.}

\item{fielddata}{A data frame containing field data of plant species
observations.}

\item{DataType}{A character string corresponding with the column in fielddata
data frame which contains the methodology used for the observation. Column
data must contain either "COVER" or "LPI". If an observation is type
"COVER" then the the function will use the 'Prcnt_C' value to populate the
response variable if the selected taxa is present at the location. If an
observation is type "LPI", then the function will give the response
variable a value of 1 if the selected taxa is present at the location. In
either case, locations where the selected taxa is absent are assigned a
value of 0 for the response variable. Default is 'DataType'.}

\item{Prcnt_C}{A character string corresponding to the column in the
fielddata data frame which contains the value for percent cover (if
applicable). Default is 'Prcnt_C'.}
}
\value{
Returns a list of five data frames: LpiCov, ObsData_plots,
ObsData_pts, PLOTS, and PTS. The function 'fsvm_train()' uses the outputs
'PLOTS' AND 'PTS' for training the machine learning models. 'LpiCov'
contains fielddata with a populated 'ResponseVariable' column based on the
provided taxonomy inforation and 'DataType'. 'ObsData_plots' and
'ObsData_pts' correspond respectively with 'LpiCov' split by 'COVER' and
'LPI' DataTypes. 'PLOTS' and 'PTS' are the preppared data frames for
machine learning analysis with 'PLOTS' containing 'Total', 'Prop', and
'Present' columns while 'PTS' contains 'Total','Hit', 'Present', and 'Prop'
columns.
}
\description{
Creates a response variable for machine learning models based on whether data
was derrived from cover plot methodology or line point intercept methodology.
The field data frame should contain a column indicating methodology ("COVER"
or "LPI") used to obtain each observation. This column name should be
provided to the "DataType" parameter. Observations obtained using the "COVER"
methodology should also contain values in a separate column corresponding
with percent cover. The column name containing this information should be
provdied to the "Prcnt_C" parameter.
}
\examples{
# # NOT RUN:
> fsvm.dat <- fsvm::as_fsvm(genus = "Abies",
                            species = "lasiocarpa",
                            subspecies = NULL,
                            fielddata = fielddata,
                            DataType = "DataType",
                            Prcnt_C = "Prcnt_C"
# # END NOT RUN

}
