% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_ecoregions.R
\name{assign_ecoregions}
\alias{assign_ecoregions}
\title{Assign Bailey's Eco-Region to field data}
\usage{
assign_ecoregions(DAT = field_data)
}
\arguments{
\item{DAT}{A data frame containing at minimum two columns named "Easting" and
"Northing". These are required for coercing the field data into a
SpatialPointsDataFrame (package::sp) so a geometric intersection can be
performed. Best behavior if coordinates already in Idaho Transverse Mercator
projection.}
}
\value{
Returns a new data frame containing original data provided with the
addition of an 'ECOCODE' column containing a string corresponding to an
eco-region (Bailey's) for each coordinate.
}
\description{
Converts field data into a spatial points data frame and assigns eco-region
based on intersection with an eco-region shapefile (internal). Uses Idaho
Transverse Mercator projection.
}
\examples{
# # NOT RUN:
>head(FieldDataPoints_QuadPolyID)
  QuadPoly_ID     quad PlotKey TranKey Easting Northing Shape_Length Shape_Area
q46115b6_6998  q46115b6     06A     06A 2367255 1671374         2688       6106
q46115b6_14717 q46115b6     10A     10A 2367548 1669244         5636      14964
q46115b8_46652 q46115b8  MFCL03  MFCL03 2349138 1661092         1040       3034

>EcoCodes<-assign_ecoregions(FieldDataPoints_QuadPolyID)
>head(EcoCodes)
  QuadPoly_ID     quad PlotKey TranKey Easting Northing Shape_Length Shape_Area ECOCODE
q46115b6_6998  q46115b6     06A     06A 2367255 1671374         2688       6106   M333D
q46115b6_14717 q46115b6     10A     10A 2367548 1669244         5636      14964   M333D
q46115b8_46652 q46115b8  MFCL03  MFCL03 2349138 1661092         1040       3034   M333D

save(EcoCodes,file=here(file.path("data","EcoCodes.RData")))
# # END NOT RUN
}
