% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ForageSpecies}
\alias{ForageSpecies}
\title{Forage utility of plant species for target game species.}
\format{
A data frame with 3142 rows and 8 variables
\describe{
\item{Genus.corrected}{corrected plant genus name}
\item{species.corrected}{corrected plant species name, specific epithet only}
\item{subspecies.variant.corrected}{corrected plant subspecies or variant name, infraspecific epithet only}
\item{SpName}{full taxonomic name of plant species, Genus specific epithet infraspecific epithet}
\item{MuleDeer}{whether mule deer use the plant taxa as forage, Y or blank}
\item{SageGrouse}{whether sage grouse use the plant taxa as forage, Y or blank}
\item{Elk}{whether elk use the plant taxa as forage, Y or blank}
\item{Moose}{whether moose use the plant taxa as forage, Y or blank}
}
}
\source{
{Idaho Department of Fish and Game}
}
\usage{
ForageSpecies
}
\description{
A dataset containing the utility of over 3,000 Idaho plant species to four
important game species.
}
\keyword{datasets}
