% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsvm_predict.R
\name{fsvm_predict}
\alias{fsvm_predict}
\title{Get predictions from trained machine learning models}
\usage{
fsvm_predict(
  pred_data = pred_data,
  pred_data_ids = pred_data_ids,
  mod_file = mod_file,
  miss_rows = miss_rows,
  mod_type = c("presence", "percent_cover")
)
}
\arguments{
\item{pred_data}{A data frame produced from 'prep_pred_covs()' which contains
the covariate data for the desired extent.}

\item{pred_data_ids}{A data frame produced from 'prep_pred_covs()' which
contains the quadpolygon IDs corresponding with the covariates.}

\item{mod_file}{A character sting corresponding to the file path containing a
model output from 'fsvm_train()' and 'get_forage_models()'. File is a
*.RData corresponding to a modeled taxonomic group
(ex."Genus_species_subspecies.RData") and contains two objects: "DAT" and
"RES" which contain the training data and results of model training
respectively.}

\item{miss_rows}{A data frame produced from 'prep_pred_covs()' which contains
all the 'NA' observations of the covariates.}

\item{mod_type}{A character string corresponding with modeling type
("presence" or "percent_cover").}
}
\value{
A list containing prediction model ('pred_model'), the prediction
('pred') and a character string of the best model ('best_model').
}
\description{
Selects best model from each output folder of 'fsvm_train()' and refits it
with the entire data set. Function uses output list from 'prep_pred_covs()'
}
\seealso{
'fsvm_train', 'get_forage_models', and 'prep_pred_covs'
}
