% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsvm_train.R
\name{fsvm_train}
\alias{fsvm_train}
\title{Train fine scale vegetation models with machine learning}
\usage{
fsvm_train(DAT = DAT, type = c("presence", "percent_cover"))
}
\arguments{
\item{DAT}{A data frame of prepared field data ('fsvm' object)}

\item{type}{A character string corresponding with model type. "presence"
corresponds with presence/absence oriented machine learning models and
"percent_cover" corresponds with proportion oriented machine learning
models.}
}
\value{
A list containing five lists: DATA, CONTROL, MODELS, COMPARE, and
BEST. DATA contains the training data set, CONTROL contains the definitions
of how the models were defined to function, MODELS contains the list of
machine learning models fitted, COMPARE contains the model summaries and
plots, and BEST contains the best fitted model
}
\description{
Trains the machine learning models on the provided formatted field data.
Presence data is trained on GLM, GBM, GLMNET, GLMNET_class, RF, and XGBOOST.
Percent cover data is trained on LM, GLM, GBM, GLMNET, RF, XGBOOST, RPART,
RPART2, TREEBAG, and BAGEARTH. This function relies on the 'caret' package.
}
\examples{
# # NOT RUN:
>presence_train.fsvm <- fsvm_train(DAT = fielddata.fsvm, type = "presence")
>view(presence_train.fsvm)

# # END NOT RUN
}
