% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_extent_shp.R
\name{get_extent_shp}
\alias{get_extent_shp}
\title{Get spatial extent as a shapefile}
\usage{
get_extent_shp(data, x, y, crs = idtm, filepath, shp.name)
}
\arguments{
\item{data}{A data frame or 'sf' object containing coordinates of field data.}

\item{x}{Column of data frame corresponding with x coordinate.}

\item{y}{Column of data frame corresponding with y coordinate.}

\item{crs}{A proj4string corresponding with current projection of data frame
coordinates. Default is Idaho Transverse Mercator ('idtm'). Not required if
using an 'sf' object has a defined projection.}

\item{filepath}{A character string leading to the output location in which to
save the output ESRI shapefile.}

\item{shp.name}{A character string corresponding with the file name of the
output shapefile.}
}
\value{
A character string corresponding with the file path of the created
shapefile.
}
\description{
Extracts the spatial extent of field data and exports as a shapefile. Meant
as a helper function for py_extract_quadpolyID() to make covariate extraction
easier for model predictions.
}
\examples{
# # NOT RUN:
>cpnwh.extent<-get_extent_shp(data=sf.cpnwh,x=NULL,y=NULL,crs=NULL,
                            filepath = "C:/Users/rritson/Documents/GitHub/HerbariumData",
                            shp.name = "cpnwh_extent")
>cpnwh.extent
[1] "C:/Users/rritson/Documents/GitHub/HerbariumData/cpnwh_extent.shp"
# # END NOT RUN

}
