% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_forage_models.R
\name{get_forage_models}
\alias{get_forage_models}
\title{Get forage models from field data}
\usage{
get_forage_models(
  fielddata = fielddata,
  FieldDataPoints_QuadPolyID = FieldDataPoints_QuadPolyID,
  Covariates_QuadPolyID = Covariates_QuadPolyID,
  EcoCodes = NULL,
  scientific.name = "ModelCode",
  model = c("Genus", "Species", "SubSpecies"),
  type = c("presence", "percent_cover")
)
}
\arguments{
\item{fielddata}{A data frame containing field data}

\item{FieldDataPoints_QuadPolyID}{A data frame containing field data points
with quadpolygon ID extracted}

\item{Covariates_QuadPolyID}{A data frame containing covariates to field data
with quadpolygon ID extracted}

\item{EcoCodes}{A data frame containing field data points with assigned
ecoregions. Default (NULL) prompts function to execute
'assign_ecoregions()' on 'FieldDataPoints_QuadPolyID'.}

\item{scientific.name}{A character string of the column name in 'fielddata'
which contains the full taxonomic name. Default is 'ModelCode'.}

\item{model}{A character string corresponding with modeling group you want to
evaluate. Must be one of "Genus", "Species", or "SubSpecies". This
selection also prompts which forage model data frame to use (a system file
generated from 'ForageSpecies' data). This parameter also populates the
subfolder name of the output.}

\item{type}{A character string corresponding with the type of model you want
to execute. Must be one of "presence" or "percent_cover". This also
generates a subfolder name to store the output (within subfolder created by
'model' parameter).}
}
\value{
Model outputs are saved to the folder path defined as
here(file.path("models","forage.model","type","file.RDATA")) where
"forage.model" corresponds with the provided 'model' parameter, "type"
corresponds with the provided 'type' parameter and "file.RData"
is generated from the taxonomic group modeled
(ex."Genus_species_subspecies.RData").
}
\description{
Formats field data and trains machine learning forage models. Console menu
prompts selection of modeling category (Genus, Species, or Subspecies) as
well as model type (Presence or Percent Cover). Function performs data
formatting with 'as_fsvm()', trains models with 'fsvm_train()', and exports
the results into appropriate folder from 'here()' file path. If approriate
folder names don't exist on the 'here()' path, then folders are created with
'dir.create()'.
}
\examples{
# # NOT RUN:
>get_forage_models(fielddata = fielddata,
                   FieldDataPoints_QuadPolyID = FieldDataPoints_QuadPolyID,
                   Covariates_QuadPolyID = Covariates_QuadPolyID,
                   EcoCodes = EcoCodes,
                   scientific.name = "ModelCode",
                   model = "Genus",
                   type = "presence")

# # END NOT RUN
}
