% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCentroidMap.R
\name{getCentroidMap}
\alias{getCentroidMap}
\title{Get centroid maps of model predictions by spatial extent}
\usage{
getCentroidMap(
  pred_covs,
  folder = c("genus", "species", "subspecies"),
  type = c("presence", "percent_cover"),
  title = "Name of Spatial Extent",
  files = NULL
)
}
\arguments{
\item{pred_covs}{A character string corresponding with the file path for an
output from 'prep_pred_covs()'.}

\item{folder}{Leave default 'c("genus","species","subspecies")' in order to
loop through all the model groups. If a different loop length is desired,
then can define any of the options as single character string.}

\item{type}{Leave default 'c("presence","perecent_cover") in order to loop
through all model types. If a different loop length is desired, then can
define either as a single character string.}

\item{title}{A character sting corresponding with the name of the spatial
extent used to create 'pred_covs' parameter.}

\item{files}{Leave default of 'NULL' in order to loop through all available
files for a model group folder. Define a vector of character strings here
if a different loop length is desired.}
}
\value{
Maps are saved as *.tif files to the folder path
here(file.path("results",folder[\link{f}],type[\link{t}],"maps",paste(title,".tif",sep=""))).
}
\description{
Creates maps of model predictions by defined spatial extent
}
\seealso{
'prep_pred_covs','getSummary'
}
