% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPredictions.R
\name{getPredictions}
\alias{getPredictions}
\title{Get predictions from all models}
\usage{
getPredictions(
  pred_covs = pred_covs,
  title = "Name of a Spatial Extent",
  filepath = here::here(),
  mod_types = c("presence", "percent_cover"),
  folders = c("genus", "species", "subspecies"),
  mod_files = NULL
)
}
\arguments{
\item{pred_covs}{A character string corresponding with the file path for the
output from 'prep_pred_covs()'.}

\item{title}{A character string corresponding with the name of the spatial
extent used for 'prep_pred_covs()'.}

\item{filepath}{A character string defining 'here::set_here(path =
filepath)'. This is used to define the file.path() of the folder containing
the trained models and destination of the prediction results. Default is
'here::here()' which is the system's current working directory.}

\item{mod_types}{Leave default 'c("presence","perecent_cover") in order to
loop through all model types. If a different loop length is desired, then
can define either as a single character string.}

\item{folders}{Leave default 'c("genus","species","subspecies")' in order to
loop through all the model groups. If a different loop length is desired,
then can define any of the options as single character string.}

\item{mod_files}{Leave default of 'NULL' in order to loop through all
available files for a model group folder. Define a vector of character
strings here if a different loop length is desired.}
}
\value{
Model predictions are saved to the folder path
here(file.path("results",folder,mod_type,paste(title,"_",mod_files\link{k},".csv")).
}
\description{
Uses parallel processing framework to loop through all model outputs (from
'fsvm_train') with 'fsvm_predict'. Iterates predictions by model type
("presence" or "percent_cover"), folders (modeling group; "genus", "species"
or "subspecies"), and model files (ex. "Genus_species_subspecies.RData"). If
this too slow or ineffective, may need to examine additional batch processing
or command-line options.
}
\seealso{
'pred_prep_covs', 'fsvm_predict', 'fsvm_train'
}
