% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSummary.R
\name{getSummary}
\alias{getSummary}
\title{Get summary of FSVM predictions}
\usage{
getSummary(
  folders = c("genus", "species", "subspecies"),
  mod_types = c("presence", "percent_cover"),
  mod_files = NULL
)
}
\arguments{
\item{folders}{A character string or vector of character strings
corresponding with model groups. Default iterates through all three
modeling groups (genus, species, and subspecies).}

\item{mod_types}{A character string or vector of character strings
corresponding to model types. Default interates through both presence and
percent cover models.}

\item{mod_files}{A character string or vector of character strings
corresponding with full file names of outputs from 'getPredictions'.
Default (NULL) triggers loop for all files within destination path
here(file.path("models",folder,type)).}
}
\value{
A summary file is saved to the path
here("results",file.path(paste(type,"_summary_trained_models.RData")))
which contains model summaries for genus, species, and subspecies.
}
\description{
Iterates through output files from 'GetPredictions' to report model
summaries. Parameters can be adjusted to restrict loop lengths.
}
\seealso{
'getPredictions'
}
