% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSVS_beta.R
\name{getSVS}
\alias{getSVS}
\title{Download Simple Veg Survey data from Survey123}
\usage{
getSVS(
  py.path = "C:/Python27/ArcGIS10.6/python.exe",
  featureService_ID,
  output_format = "CSV",
  download_folder = "C:/Temp/",
  agol.username,
  agol.password,
  download_url,
  filename,
  token,
  folder
)
}
\arguments{
\item{py.path}{Character string of file path to a valid installation of
Python. If using default path, must be running R in 32-bit session with a
valid installation of ArcGIS.}

\item{featureService_ID}{A character string containing a valid feature
service id.}

\item{output_format}{A character string corresponding with the desired output
format. Must be either "CSV", "Shapefile", or "File Geodatabase".
Recommended to leave at default of "CSV" in order for returned data to be
properly formatted. Other selections will be returned as is.}

\item{agol.username}{A character string containing a valid ArcGIS Online IDFG
account (firstname.lastname_IDFGgis) with AGOL license.}

\item{agol.password}{A character string containing a valid password
corresponding to specified AGOL username.}

\item{download_url}{A character string containing the URL to the desired
survey.}

\item{filename}{A character string corresponding with the desired filename of
the downloaded data.}

\item{token}{A character string containing a valid site token for downloading
the survey.}

\item{folder}{A character string corresponding with the desired folder to
contain the downloaded data.}
}
\value{
A data frame containing the formatted output (if output_format is
"CSV") or as-is downloaded data file (shp or gdb).
}
\description{
Accesses ArcGIS Online (AGOL) for a permitted IDFG account user with a python
script (downloadSVS.py) which downloads data from a specified survey.
Requires 32-bit R unless a valid alternative py.path is specified (may be
able to use default available Python with 64-bit). A valid AGOL username and
password, the featureServiceID, the url of the survey, and download token
must all be specified. Also requires a file path to a download folder and
filename. Output format is recommended to remain "CSV", however "Shapefile"
and "File Geodatabase" are valid alternatives. Only an output parameter of
"CSV" will trigger the function to format the output, other formats will be
returned as is.
}
