% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_mod_covs.R
\name{prep_mod_covs}
\alias{prep_mod_covs}
\title{Prepare Covariates for Model Training}
\usage{
prep_mod_covs(
  covariates = Covariates_QuadPolyID,
  field.pts = FieldDataPoints_QuadPolyID
)
}
\arguments{
\item{covariates}{A data frame containing covariates corresponding with field
data points with extracted quadpolygon ids (an output of
'py_extract_quadpolyID()').}

\item{field.pts}{A data frame containing field data points with extracted
quadpolygon ids (an output of 'py_extract_quadpolyID()').}
}
\value{
A formatted version of the data frame provided to the 'covariates'
parameter.
}
\description{
Formats covariates extracted for field data for training machine learning
models with 'fsvm_train()'
}
