% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_pred_covs.R
\name{prep_pred_covs}
\alias{prep_pred_covs}
\title{Prepare Covariates for Model Predictions}
\usage{
prep_pred_covs(
  filepath = filepath,
  cov_files = cov_files,
  filetype = c("dbf", "csv")
)
}
\arguments{
\item{filepath}{A character string corresponding with the file path to the
output location of the covariate files.}

\item{cov_files}{A character string of the file name or list of file names
corresponding with an output covariate file.}

\item{filetype}{The file type of the output table. Use 'dbf' if table is a
merged dbf creation and 'csv' if output table is a full 100k polygon set.}
}
\value{
A list of three data frames: 'pred_data', 'pred_data_ids', and
'miss_rows'. 'pred_data' contains the covariate data to be used in model
predictions 'pred_data_ids' contains the QuadPolyIDs corresponding with
prediction covariates. 'miss_rows' contains all observations of 'NA'.
}
\description{
Formats covariates extracted for the entire spatial extent of the field data
for use in making predictions from the trained machine learning models
}
