% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{quads_100k}
\alias{quads_100k}
\title{Names of eCognition 100k quadpolygons.}
\format{
A vector with 75 elements containing character strings
}
\source{
{Idaho Department of Fish and Game}
}
\usage{
quads_100k
}
\description{
A vector containing the names of the eCognition 100k quadpolygons located
within the file geodatases.
}
\keyword{datasets}
