% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectify_taxa.R
\name{rectify_taxa}
\alias{rectify_taxa}
\title{Rectify taxonomic names and assign modeling groups}
\usage{
rectify_taxa(dat, scientific.name = "ModelCode")
}
\arguments{
\item{dat}{A data frame with at least one column containing taxonomic names
of plants}

\item{scientific.name="ModelCode"}{A character string corresponding with the
column in the data frame which contains the taxonomic plant names. Default
("ModelCode") corresponds with the appropriate column in the current field
data file.}
}
\value{
Returns the input data frame appended with 8 additional columns:
"G1", "G2", "G3", "G4", "G1.TSN", "G2.TSN", "G3.TSN", and "G4.TSN". "G1"
corresponds with the accepted genera name and "G1.TSN" the corresponding
taxonomic serial number. "G2" corresponds with species (Genus and specific
epithet only), "G3" correspons with subspecies only (Genus specific
eptithet ssp. subspecies), and "G4" corresponds with varieties only (Genus
specific epithet var. variety). Note, for example that "G2" and "G2.TSN"
would be identicial for subspecies and varieties with the same taxonomic
parent.
}
\description{
Uses Integrated Taxonomic Information System to assign current accepted name
and taxonomic serial number (TSN) to field data by modeling group
("G1","G2","G3",and "G4").
}
\examples{
# # NOT RUN:
> foo<- fielddata[grepl("Artemisia",fielddata$ModelCode),]
> foo[c("42","1575","132185"),c(1,3:7,9,11)]
              TranKey     Source  DataType Smpl_Yr Easting Northing    Code                               ModelCode
IDFG_2019_q42112c8_5750  IDFG_LPI      LPI    2019  2584333  1239448   ARTR2                    Artemisia tridentata
                 Agnes    BLM_SC      LPI    2015  2672930  1436288   ARTRV      Artemisia tridentata ssp. vaseyana
            1.40902E+15 DIMA_Lem      LPI    2014  2522934  1554701  ARTRW8  Artemisia tridentata ssp. wyomingensis

>foo<-rectify_taxa(dat = fielddata, scientific.name="ModelCode")
>foo[c("42","1575","132185"),c(1,17:25)]
                      TranKey                          Accepted.Name        G1                   G2                                     G3 G4
42     IDFG_2019_q42112c8_5750                   Artemisia tridentata Artemisia Artemisia tridentata                                     NA NA
1575                     Agnes     Artemisia tridentata ssp. vaseyana Artemisia Artemisia tridentata     Artemisia tridentata ssp. vaseyana NA
132185             1.40902E+15 Artemisia tridentata ssp. wyomingensis Artemisia Artemisia tridentata Artemisia tridentata ssp. wyomingensis NA
      G1.TSN G2.TSN G3.TSN G4.TSN
42      35431  35498   <NA>   <NA>
1575    35431  35498 183740   <NA>
132185  35431  35498 183741   <NA>

##Note differences!
# # END NOT RUN

}
