% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_extent_manual.R
\name{set_extent_manual}
\alias{set_extent_manual}
\title{Set Manual Spatial Extent}
\usage{
set_extent_manual(
  extent = c("Region", "GMU", "UID_100k", "UID_24k"),
  selections = choices
)
}
\arguments{
\item{extent}{A character string corresponding with the extent type you want
to choose from. Choices are "Region", "GMU", "UID_110k", or "UID_24k".
Multiple choices can be selected and stored as character vector, however
this is probably less useful.}

\item{selections}{A character vector corresponding with the selection(s)
within the extent type.}
}
\value{
A character vector containing the names of 100k quadpolygons
corresponding with the selected location extents. These name should be able
to be passed to 'py_extract_quadpolyID' as quad selections.
}
\description{
Returns a list of quadpolygons which correspond with a selected extent. This
includes choices of specific IDFG Regions (1-7), IDFG Game Management Units,
or UIDs of 100k or 24k quadpolygons.
}
\examples{
# # # NOT RUN:
# A single GMU
> gmu1.quadpolys <-set_extent_manual(extent = "GMU", selections = "1")
> gmu1.quadpolys
[1] "q48117e1_100blk" "q48116e1_100blk" "q48117a1_100blk" "q48116a1_100blk"

# A single Region
> r7.quadpolys <-set_extent_manual(extent = "Region", selections = "7")
> r7.quadpolys
[1] "q46115a1_100blk" "q46114a1_100blk" "q45115e1_100blk" "q45114e1_100blk" "q45114a1_100blk" "q45115a1_100blk" "q45113e1_100blk" "q45113a1_100blk" "q44115e1_100blk"
[10] "q44115a1_100blk" "q44114e1_100blk" "q44114a1_100blk" "q44113e1_100blk" "q44113a1_100blk" "q44112a1_100blk" "q43115e1_100blk" "q43114e1_100blk" "q44116a1_100blk"
[19] "q43116e1_100blk" "q43116a1_100blk" "q43115a1_100blk" "q43114a1_100blk" "q43113e1_100blk" "q43113a1_100blk" "q43112e1_100blk"

# Multiple GMUs
> eoe.quadpolys <-set_extent_manual(extent = "GMU", selections = c("1","6","10A"))
> eoe.quadpolys
[1] "q48117e1_100blk" "q48116e1_100blk" "q48117a1_100blk" "q48116a1_100blk" "q47116a1_100blk" "q47115a1_100blk" "q46116e1_100blk" "q46115e1_100blk" "q46116a1_100blk"
[10] "q46115a1_100blk"
# # # END NOT RUN # # #

}
\seealso{
py_extract_quadpolyID
}
